rename use-agents -> use-agent, final
