+++ docs/reference/use-agent.mdx
@@ -3,10 +3,10 @@ title: useAgents
 description: 'React hook for building real-time, multi-threaded AI applications'
 ---
 
-The `useAgents` hook is the core of the `@inngest/use-agents` package. It's a comprehensive, client-side hook for React that manages real-time, multi-threaded conversations with an AgentKit network. It encapsulates the entire lifecycle of agent interactions, including sending messages, receiving streaming events, handling out-of-order event sequences, and managing connection state.
+The `useAgents` hook is the core of the `@inngest/use-agent` package. It's a comprehensive, client-side hook for React that manages real-time, multi-threaded conversations with an AgentKit network. It encapsulates the entire lifecycle of agent interactions, including sending messages, receiving streaming events, handling out-of-order event sequences, and managing connection state.
 
 ```tsx
-import { useAgents, AgentProvider } from '@inngest/use-agents';
+import { useAgents, AgentProvider } from '@inngest/use-agent';
 
 function App() {
   return (
