+++ docs/streaming/events.mdx
@@ -2,7 +2,7 @@
 title: Events
 ---
 
-The `use-agents` hook is built on an event-driven architecture. Real-time events are streamed from the server, processed by a state reducer, and used to build the conversation UI incrementally. This document details the events, their purpose, and how you can use them.
+The `use-agent` hook is built on an event-driven architecture. Real-time events are streamed from the server, processed by a state reducer, and used to build the conversation UI incrementally. This document details the events, their purpose, and how you can use them.
 
 ## Understanding the Event Lifecycle
