+++ docs/streaming/transport.mdx
@@ -75,7 +75,7 @@ If you provide a partial configuration, it will be merged with the default trans
 ### Example: Customizing API routes, headers, and body
 
 ```tsx
-import { AgentProvider } from "@inngest/use-agents";
+import { AgentProvider } from "@inngest/use-agent";
 
 function App({ userId, getAuthToken, getTenantId }) {
   return (
