+++ examples/use-agent/components/chat/message-parts/TextMessagePart.tsx
@@ -1,5 +1,5 @@
 
-import type { TextUIPart } from "@inngest/use-agents";
+import type { TextUIPart } from "@inngest/use-agent";
 import { Response } from '@/components/ai-elements/response';
 
 interface TextMessagePartProps {
