+++ examples/use-agent/hooks/index.ts
@@ -1,5 +1,5 @@
 // UI Helper Hooks for the use-agent example application
-// These hooks are specific to this UI implementation and are not part of the core use-agents package
+// These hooks are specific to this UI implementation and are not part of the core use-agent package
 
 export { useEditMessage } from './use-edit-message';
 export { useMessageActions } from './use-message-actions';
