+++ examples/use-agent/inngest/client.ts
@@ -2,6 +2,6 @@ import { realtimeMiddleware } from '@inngest/realtime/middleware';
 import { Inngest } from 'inngest';
 
 export const inngest = new Inngest({
-  id: 'use-agents-demo-client',
+  id: 'use-agent-demo-client',
   middleware: [realtimeMiddleware()],
 });
