+++ examples/use-agent/lib/use-insights-agent.ts
@@ -4,7 +4,7 @@ import {
   useAgents,
   type UseAgentsConfig,
   type UseAgentsReturn,
-} from "@inngest/use-agents";
+} from "@inngest/use-agent";
 
 // Import the ToolManifest built from server-side tool definitions
 import type { ToolManifest } from "@/app/api/inngest/functions/agents/types";
