+++ packages/use-agent/package.json
@@ -39,7 +39,7 @@
   "repository": {
     "type": "git",
     "url": "git+https://github.com/inngest/agent-kit.git",
-    "directory": "packages/use-agents"
+    "directory": "packages/use-agent"
   },
   "peerDependencies": {
     "@inngest/realtime": ">=0.3.0",
